feat: minor setup refactor + sync opened thread + opened models
