+++ packages/frontend-2/pages/projects/[id]/models/[modelId]/index.vue
@@ -1,62 +1,64 @@
 <template>
-  <div class="absolute top-0 left-0 w-screen h-screen">
-    <!-- Nav -->
-    <Portal to="navigation">
-      <ViewerScope :state="state">
-        <HeaderNavLink
-          :to="`/projects/${project?.id}`"
-          :name="project?.name"
-        ></HeaderNavLink>
-        <ViewerExplorerNavbarLink />
-      </ViewerScope>
-    </Portal>
+  <ViewerPostSetupWrapper>
+    <div class="absolute top-0 left-0 w-screen h-screen">
+      <!-- Nav -->
+      <Portal to="navigation">
+        <ViewerScope :state="state">
+          <HeaderNavLink
+            :to="`/projects/${project?.id}`"
+            :name="project?.name"
+          ></HeaderNavLink>
+          <ViewerExplorerNavbarLink />
+        </ViewerScope>
+      </Portal>
 
-    <Portal to="primary-actions">
-      <div class="flex space-x-4">
-        <FormButton :icon-left="ShareIcon">Share</FormButton>
-      </div>
-    </Portal>
+      <Portal to="primary-actions">
+        <div class="flex space-x-4">
+          <FormButton :icon-left="ShareIcon">Share</FormButton>
+        </div>
+      </Portal>
 
-    <ClientOnly>
-      <!-- Tour host -->
-      <div
-        v-if="tourState.showTour"
-        class="z-30 fixed w-full h-full flex justify-center items-center pointer-events-none"
-      >
-        <TourOnboarding />
-      </div>
-      <!-- Viewer host -->
-      <div class="special-gradient absolute w-screen h-screen z-10">
-        <ViewerBase />
+      <ClientOnly>
+        <!-- Tour host -->
+        <div
+          v-if="tourState.showTour"
+          class="z-30 fixed w-full h-full flex justify-center items-center pointer-events-none"
+        >
+          <TourOnboarding />
+        </div>
+        <!-- Viewer host -->
+        <div class="special-gradient absolute w-screen h-screen z-10">
+          <ViewerBase />
+          <Transition
+            enter-from-class="opacity-0"
+            enter-active-class="transition duration-1000"
+          >
+            <ViewerAnchoredPoints v-show="tourState.showViewerControls" />
+          </Transition>
+        </div>
+
+        <!-- Global loading bar -->
+        <ViewerLoadingBar class="z-20" />
+
+        <!-- Sidebar sketches -->
         <Transition
           enter-from-class="opacity-0"
           enter-active-class="transition duration-1000"
         >
-          <ViewerAnchoredPoints v-show="tourState.showViewerControls" />
+          <ViewerControls v-show="tourState.showViewerControls" class="z-20" />
         </Transition>
-      </div>
-
-      <!-- Global loading bar -->
-      <ViewerLoadingBar class="z-20" />
-
-      <!-- Sidebar sketches -->
-      <Transition
-        enter-from-class="opacity-0"
-        enter-active-class="transition duration-1000"
-      >
-        <ViewerControls v-show="tourState.showViewerControls" class="z-20" />
-      </Transition>
-      <!-- Viewer Object Selection Info Display -->
-      <Transition
-        enter-from-class="opacity-0"
-        enter-active-class="transition duration-1000"
-      >
-        <ViewerSelectionSidebar v-show="tourState.showViewerControls" class="z-20" />
-      </Transition>
-      <!-- Shows up when filters are applied for an easy return to normality -->
-      <ViewerGlobalFilterReset class="z-20" />
-    </ClientOnly>
-  </div>
+        <!-- Viewer Object Selection Info Display -->
+        <Transition
+          enter-from-class="opacity-0"
+          enter-active-class="transition duration-1000"
+        >
+          <ViewerSelectionSidebar v-show="tourState.showViewerControls" class="z-20" />
+        </Transition>
+        <!-- Shows up when filters are applied for an easy return to normality -->
+        <ViewerGlobalFilterReset class="z-20" />
+      </ClientOnly>
+    </div>
+  </ViewerPostSetupWrapper>
 </template>
 <script setup lang="ts">
 import { graphql } from '~~/lib/common/generated/gql'
